+++ CHANGELOG.markdown
@@ -3,6 +3,11 @@ next
 * Introduce the `Data.Functor.Classes.Generic` module, which provides functions that can generically implement methods in the `Eq1`, `Ord1`, `Read1`, and `Show1` classes (without the usual boilerplate involved).
 * Introduce the `generic-deriving` flag. When enabled, `transformers-compat` will depend on the `generic-deriving` library on older versions of GHC to backport `GHC.Generics` support for `Generic` instances and the machinery in `Data.Functor.Classes.Generic`.
 * Some instances were present in `Data.Functor.Sum` but not in `Control.Monad.Trans.Instances` (e.g., the `Generic`, `Typeable`, and `Data` instances for `Sum`). This has been fixed.
+* Backport changes from `transformers-0.5.5`:
+  * Backport the `Semigroup` instance for `Constant`
+  * Add `mapSelect` and `mapSelectT`
+  * Define `selectToContT` (and deprecate `selectToCont` in favor of that)
+  * Define some explicit `(*>)` definitions to avoid space leaks
 * Backport changes from `transformers-0.5.4` (i.e., add `Bifoldable` and `Bitraversable` instances for `Data.Functor.Constant`)
 * Backport changes from `transformers-0.5.3`:
   * Backport the `Control.Monad.Trans.Accum` and `Control.Monad.Trans.Select` modules